Moved shared AWS constants & functions to awscommon
